+++ src/modules/edc-demo/components/asset-editor-dialog/model/http-params-form-auth-header-type.ts
@@ -1,1 +1,0 @@
-export type HttpParamsFormAuthHeaderType = 'None' | 'Value' | 'Vault-Secret';